fix: rename RealtimeVoteCounting to VoteCounting
