+++ src/components/HomeView/VoteCounting.vue
